Extract provider model catalog helpers
